Crate rusoto_route53

Amazon Route 53 is a highly available and scalable Domain Name System (DNS) web service.

If you're using the service, you're probably looking for Route53Client and Route53.

Modules

util
Utility helpers for working with Route53

Structs

AccountLimit
A complex type that contains the type of limit that you specified in the request and the current value for that limit.

AccountLimitTypeSerializer

AlarmIdentifier
A complex type that identifies the CloudWatch alarm that you want Amazon Route 53 health checkers to use to determine whether the specified health check is healthy.

AlarmIdentifierSerializer

AlarmNameSerializer

AliasHealthEnabledSerializer

AliasTarget

Alias resource record sets only: Information about the AWS resource, such as a CloudFront distribution or an Amazon S3 bucket, that you want to route traffic to.

When creating resource record sets for a private hosted zone, note the following:

Creating geolocation alias resource record sets or latency alias resource record sets in a private hosted zone is unsupported.

For information about creating failover resource record sets in a private hosted zone, see Configuring Failover in a Private Hosted Zone.

AliasTargetSerializer

AssociateVPCCommentSerializer

AssociateVPCWithHostedZoneRequest
A complex type that contains information about the request to associate a VPC with a private hosted zone.

AssociateVPCWithHostedZoneRequestSerializer

AssociateVPCWithHostedZoneResponse
A complex type that contains the response information for the AssociateVPCWithHostedZone request.

Change
The information for each resource record set that you want to change.

ChangeActionSerializer

ChangeBatch
The information for a change request.

ChangeBatchSerializer

ChangeInfo
A complex type that describes change information about changes made to your hosted zone.

ChangeResourceRecordSetsRequest
A complex type that contains change information for the resource record set.

ChangeResourceRecordSetsRequestSerializer

ChangeResourceRecordSetsResponse
A complex type containing the response for the request.

ChangeSerializer

ChangeTagsForResourceRequest
A complex type that contains information about the tags that you want to add, edit, or delete.

ChangeTagsForResourceRequestSerializer

ChangeTagsForResourceResponse
Empty response for the request.

ChangesSerializer

ChildHealthCheckListSerializer

CloudWatchAlarmConfiguration
A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

CloudWatchLogsLogGroupArnSerializer

CloudWatchRegionSerializer

CreateHealthCheckRequest
A complex type that contains the health check request information.

CreateHealthCheckRequestSerializer

CreateHealthCheckResponse
A complex type containing the response information for the new health check.

CreateHostedZoneRequest
A complex type that contains information about the request to create a public or private hosted zone.

CreateHostedZoneRequestSerializer

CreateHostedZoneResponse
A complex type containing the response information for the hosted zone.

CreateQueryLoggingConfigRequest

CreateQueryLoggingConfigRequestSerializer

CreateQueryLoggingConfigResponse

CreateReusableDelegationSetRequest

CreateReusableDelegationSetRequestSerializer

CreateReusableDelegationSetResponse

CreateTrafficPolicyInstanceRequest
A complex type that contains information about the resource record sets that you want to create based on a specified traffic policy.

CreateTrafficPolicyInstanceRequestSerializer

CreateTrafficPolicyInstanceResponse
A complex type that contains the response information for the CreateTrafficPolicyInstance request.

CreateTrafficPolicyRequest
A complex type that contains information about the traffic policy that you want to create.

CreateTrafficPolicyRequestSerializer

CreateTrafficPolicyResponse
A complex type that contains the response information for the CreateTrafficPolicy request.

CreateTrafficPolicyVersionRequest
A complex type that contains information about the traffic policy that you want to create a new version for.

CreateTrafficPolicyVersionRequestSerializer

CreateTrafficPolicyVersionResponse
A complex type that contains the response information for the CreateTrafficPolicyVersion request.

CreateVPCAssociationAuthorizationRequest
A complex type that contains information about the request to authorize associating a VPC with your private hosted zone. Authorization is only required when a private hosted zone and a VPC were created by using different accounts.

CreateVPCAssociationAuthorizationRequestSerializer

CreateVPCAssociationAuthorizationResponse
A complex type that contains the response information from a CreateVPCAssociationAuthorization request.

DNSNameSerializer

DelegationSet
A complex type that lists the name servers in a delegation set, as well as the CallerReference and the ID for the delegation set.

DeleteHealthCheckRequest
This action deletes a health check.

DeleteHealthCheckResponse
An empty element.

DeleteHostedZoneRequest
A request to delete a hosted zone.

DeleteHostedZoneResponse
A complex type that contains the response to a DeleteHostedZone request.

DeleteQueryLoggingConfigRequest

DeleteQueryLoggingConfigResponse

DeleteReusableDelegationSetRequest
A request to delete a reusable delegation set.

DeleteReusableDelegationSetResponse
An empty element.

DeleteTrafficPolicyInstanceRequest
A request to delete a specified traffic policy instance.

DeleteTrafficPolicyInstanceResponse
An empty element.

DeleteTrafficPolicyRequest
A request to delete a specified traffic policy version.

DeleteTrafficPolicyResponse
An empty element.

DeleteVPCAssociationAuthorizationRequest
A complex type that contains information about the request to remove authorization to associate a VPC that was created by one AWS account with a hosted zone that was created with a different AWS account.

DeleteVPCAssociationAuthorizationRequestSerializer

DeleteVPCAssociationAuthorizationResponse
Empty response for the request.

Dimension
For the metric that the CloudWatch alarm is associated with, a complex type that contains information about one dimension.

DisabledSerializer

DisassociateVPCCommentSerializer

DisassociateVPCFromHostedZoneRequest
A complex type that contains information about the VPC that you want to disassociate from a specified private hosted zone.

DisassociateVPCFromHostedZoneRequestSerializer

DisassociateVPCFromHostedZoneResponse
A complex type that contains the response information for the disassociate request.

EnableSNISerializer

FailureThresholdSerializer

FullyQualifiedDomainNameSerializer

GeoLocation
A complex type that contains information about a geographic location.

GeoLocationContinentCodeSerializer

GeoLocationCountryCodeSerializer

GeoLocationDetails
A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.

GeoLocationSerializer

GeoLocationSubdivisionCodeSerializer

GetAccountLimitRequest
A complex type that contains information about the request to create a hosted zone.

GetAccountLimitResponse
A complex type that contains the requested limit.

GetChangeRequest
The input for a GetChange request.

GetChangeResponse
A complex type that contains the ChangeInfo element.

GetCheckerIpRangesRequest
Empty request.

GetCheckerIpRangesResponse
A complex type that contains the CheckerIpRanges element.

GetGeoLocationRequest
A request for information about whether a specified geographic location is supported for Amazon Route 53 geolocation resource record sets.

GetGeoLocationResponse
A complex type that contains the response information for the specified geolocation code.

GetHealthCheckCountRequest
A request for the number of health checks that are associated with the current AWS account.

GetHealthCheckCountResponse
A complex type that contains the response to a GetHealthCheckCount request.

GetHealthCheckLastFailureReasonRequest
A request for the reason that a health check failed most recently.

GetHealthCheckLastFailureReasonResponse
A complex type that contains the response to a GetHealthCheckLastFailureReason request.

GetHealthCheckRequest
A request to get information about a specified health check.

GetHealthCheckResponse
A complex type that contains the response to a GetHealthCheck request.

GetHealthCheckStatusRequest
A request to get the status for a health check.

GetHealthCheckStatusResponse
A complex type that contains the response to a GetHealthCheck request.

GetHostedZoneCountRequest
A request to retrieve a count of all the hosted zones that are associated with the current AWS account.

GetHostedZoneCountResponse
A complex type that contains the response to a GetHostedZoneCount request.

GetHostedZoneLimitRequest
A complex type that contains information about the request to create a hosted zone.

GetHostedZoneLimitResponse
A complex type that contains the requested limit.

GetHostedZoneRequest
A request to get information about a specified hosted zone.

GetHostedZoneResponse
A complex type that contain the response to a GetHostedZone request.

GetQueryLoggingConfigRequest

GetQueryLoggingConfigResponse

GetReusableDelegationSetLimitRequest
A complex type that contains information about the request to create a hosted zone.

GetReusableDelegationSetLimitResponse
A complex type that contains the requested limit.

GetReusableDelegationSetRequest
A request to get information about a specified reusable delegation set.

GetReusableDelegationSetResponse
A complex type that contains the response to the GetReusableDelegationSet request.

GetTrafficPolicyInstanceCountRequest
Request to get the number of traffic policy instances that are associated with the current AWS account.

GetTrafficPolicyInstanceCountResponse
A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

GetTrafficPolicyInstanceRequest
Gets information about a specified traffic policy instance.

GetTrafficPolicyInstanceResponse
A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

GetTrafficPolicyRequest
Gets information about a specific traffic policy version.

GetTrafficPolicyResponse
A complex type that contains the response information for the request.

HealthCheck
A complex type that contains information about one health check that is associated with the current AWS account.

HealthCheckConfig
A complex type that contains information about the health check.

HealthCheckConfigSerializer

HealthCheckIdSerializer

HealthCheckNonceSerializer

HealthCheckObservation
A complex type that contains the last failure reason as reported by one Amazon Route 53 health checker.

HealthCheckRegionListSerializer

HealthCheckRegionSerializer

HealthCheckTypeSerializer

HealthCheckVersionSerializer

HealthThresholdSerializer

HostedZone
A complex type that contains general information about the hosted zone.

HostedZoneConfig
A complex type that contains an optional comment about your hosted zone. If you don't want to specify a comment, omit both the HostedZoneConfig and Comment elements.

HostedZoneConfigSerializer

HostedZoneLimit
A complex type that contains the type of limit that you specified in the request and the current value for that limit.

HostedZoneLimitTypeSerializer

IPAddressSerializer

InsufficientDataHealthStatusSerializer

InvertedSerializer

IsPrivateZoneSerializer

LinkedService
If a health check or hosted zone was created by another service, LinkedService is a complex type that describes the service that created the resource. When a resource is created by another service, you can't edit or delete it using Amazon Route 53.

ListGeoLocationsRequest
A request to get a list of geographic locations that Amazon Route 53 supports for geolocation resource record sets.

ListGeoLocationsResponse
A complex type containing the response information for the request.

ListHealthChecksRequest
A request to retrieve a list of the health checks that are associated with the current AWS account.

ListHealthChecksResponse
A complex type that contains the response to a ListHealthChecks request.

ListHostedZonesByNameRequest
Retrieves a list of the public and private hosted zones that are associated with the current AWS account in ASCII order by domain name.

ListHostedZonesByNameResponse
A complex type that contains the response information for the request.

ListHostedZonesRequest
A request to retrieve a list of the public and private hosted zones that are associated with the current AWS account.

ListHostedZonesResponse

ListQueryLoggingConfigsRequest

ListQueryLoggingConfigsResponse

ListResourceRecordSetsRequest
A request for the resource record sets that are associated with a specified hosted zone.

ListResourceRecordSetsResponse
A complex type that contains list information for the resource record set.

ListReusableDelegationSetsRequest
A request to get a list of the reusable delegation sets that are associated with the current AWS account.

ListReusableDelegationSetsResponse
A complex type that contains information about the reusable delegation sets that are associated with the current AWS account.

ListTagsForResourceRequest
A complex type containing information about a request for a list of the tags that are associated with an individual resource.

ListTagsForResourceResponse
A complex type that contains information about the health checks or hosted zones for which you want to list tags.

ListTagsForResourcesRequest
A complex type that contains information about the health checks or hosted zones for which you want to list tags.

ListTagsForResourcesRequestSerializer

ListTagsForResourcesResponse
A complex type containing tags for the specified resources.

ListTrafficPoliciesRequest
A complex type that contains the information about the request to list the traffic policies that are associated with the current AWS account.

ListTrafficPoliciesResponse
A complex type that contains the response information for the request.

ListTrafficPolicyInstancesByHostedZoneRequest
A request for the traffic policy instances that you created in a specified hosted zone.

ListTrafficPolicyInstancesByHostedZoneResponse
A complex type that contains the response information for the request.

ListTrafficPolicyInstancesByPolicyRequest
A complex type that contains the information about the request to list your traffic policy instances.

ListTrafficPolicyInstancesByPolicyResponse
A complex type that contains the response information for the request.

ListTrafficPolicyInstancesRequest
A request to get information about the traffic policy instances that you created by using the current AWS account.

ListTrafficPolicyInstancesResponse
A complex type that contains the response information for the request.

ListTrafficPolicyVersionsRequest
A complex type that contains the information about the request to list your traffic policies.

ListTrafficPolicyVersionsResponse
A complex type that contains the response information for the request.

ListVPCAssociationAuthorizationsRequest
A complex type that contains information about that can be associated with your hosted zone.

ListVPCAssociationAuthorizationsResponse
A complex type that contains the response information for the request.

MaxResultsSerializer

MeasureLatencySerializer

NonceSerializer

PageMarkerSerializer

PageMaxItemsSerializer

PaginationTokenSerializer

PortSerializer

QueryLoggingConfig
A complex type that contains information about a configuration for DNS query logging.

QueryLoggingConfigIdSerializer

RDataSerializer

RRTypeSerializer

RequestIntervalSerializer

ResettableElementNameListSerializer

ResettableElementNameSerializer

ResourceDescriptionSerializer

ResourceIdSerializer

ResourcePathSerializer

ResourceRecord

Information specific to the resource record.

If you're creating an alias resource record set, omit ResourceRecord.

ResourceRecordSerializer

ResourceRecordSet
Information about the resource record set to create or delete.

ResourceRecordSetFailoverSerializer

ResourceRecordSetIdentifierSerializer

ResourceRecordSetMultiValueAnswerSerializer

ResourceRecordSetRegionSerializer

ResourceRecordSetSerializer

ResourceRecordSetWeightSerializer

ResourceRecordsSerializer

ResourceTagSet
A complex type containing a resource and its associated tags.

ReusableDelegationSetLimit
A complex type that contains the type of limit that you specified in the request and the current value for that limit.

ReusableDelegationSetLimitTypeSerializer

Route53Client
A client for the Route 53 API.

SearchStringSerializer

StatusReport
A complex type that contains the status that one Amazon Route 53 health checker reports and the time of the health check.

SubnetMaskSerializer

TTLSerializer

Tag
A complex type that contains information about a tag that you want to add or edit for the specified health check or hosted zone.

TagKeyListSerializer

TagKeySerializer

TagListSerializer

TagResourceIdListSerializer

TagResourceIdSerializer

TagResourceTypeSerializer

TagSerializer

TagValueSerializer

TestDNSAnswerRequest
Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify the IP address of a DNS resolver, an EDNS0 client subnet IP address, and a subnet mask.

TestDNSAnswerResponse
A complex type that contains the response to a TestDNSAnswer request.

TrafficPolicy
A complex type that contains settings for a traffic policy.

TrafficPolicyCommentSerializer

TrafficPolicyDocumentSerializer

TrafficPolicyIdSerializer

TrafficPolicyInstance
A complex type that contains settings for the new traffic policy instance.

TrafficPolicyInstanceIdSerializer

TrafficPolicyNameSerializer

TrafficPolicySummary
A complex type that contains information about the latest version of one traffic policy that is associated with the current AWS account.

TrafficPolicyVersionMarkerSerializer

TrafficPolicyVersionSerializer

UpdateHealthCheckRequest
A complex type that contains information about a request to update a health check.

UpdateHealthCheckRequestSerializer

UpdateHealthCheckResponse
A complex type that contains the response to the UpdateHealthCheck request.

UpdateHostedZoneCommentRequest
A request to update the comment for a hosted zone.

UpdateHostedZoneCommentRequestSerializer

UpdateHostedZoneCommentResponse
A complex type that contains the response to the UpdateHostedZoneComment request.

UpdateTrafficPolicyCommentRequest
A complex type that contains information about the traffic policy that you want to update the comment for.

UpdateTrafficPolicyCommentRequestSerializer

UpdateTrafficPolicyCommentResponse
A complex type that contains the response information for the traffic policy.

UpdateTrafficPolicyInstanceRequest
A complex type that contains information about the resource record sets that you want to update based on a specified traffic policy instance.

UpdateTrafficPolicyInstanceRequestSerializer

UpdateTrafficPolicyInstanceResponse
A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

VPC
(Private hosted zones only) A complex type that contains information about an Amazon VPC.

VPCIdSerializer

VPCRegionSerializer

VPCSerializer

Enums

AssociateVPCWithHostedZoneError
Errors returned by AssociateVPCWithHostedZone

ChangeResourceRecordSetsError
Errors returned by ChangeResourceRecordSets

ChangeTagsForResourceError
Errors returned by ChangeTagsForResource

CreateHealthCheckError
Errors returned by CreateHealthCheck

CreateHostedZoneError
Errors returned by CreateHostedZone

CreateQueryLoggingConfigError
Errors returned by CreateQueryLoggingConfig

CreateReusableDelegationSetError
Errors returned by CreateReusableDelegationSet

CreateTrafficPolicyError
Errors returned by CreateTrafficPolicy

CreateTrafficPolicyInstanceError
Errors returned by CreateTrafficPolicyInstance

CreateTrafficPolicyVersionError
Errors returned by CreateTrafficPolicyVersion

CreateVPCAssociationAuthorizationError
Errors returned by CreateVPCAssociationAuthorization

DeleteHealthCheckError
Errors returned by DeleteHealthCheck

DeleteHostedZoneError
Errors returned by DeleteHostedZone

DeleteQueryLoggingConfigError
Errors returned by DeleteQueryLoggingConfig

DeleteReusableDelegationSetError
Errors returned by DeleteReusableDelegationSet

DeleteTrafficPolicyError
Errors returned by DeleteTrafficPolicy

DeleteTrafficPolicyInstanceError
Errors returned by DeleteTrafficPolicyInstance

DeleteVPCAssociationAuthorizationError
Errors returned by DeleteVPCAssociationAuthorization

DisassociateVPCFromHostedZoneError
Errors returned by DisassociateVPCFromHostedZone

GetAccountLimitError
Errors returned by GetAccountLimit

GetChangeError
Errors returned by GetChange

GetCheckerIpRangesError
Errors returned by GetCheckerIpRanges

GetGeoLocationError
Errors returned by GetGeoLocation

GetHealthCheckCountError
Errors returned by GetHealthCheckCount

GetHealthCheckError
Errors returned by GetHealthCheck

GetHealthCheckLastFailureReasonError
Errors returned by GetHealthCheckLastFailureReason

GetHealthCheckStatusError
Errors returned by GetHealthCheckStatus

GetHostedZoneCountError
Errors returned by GetHostedZoneCount

GetHostedZoneError
Errors returned by GetHostedZone

GetHostedZoneLimitError
Errors returned by GetHostedZoneLimit

GetQueryLoggingConfigError
Errors returned by GetQueryLoggingConfig

GetReusableDelegationSetError
Errors returned by GetReusableDelegationSet

GetReusableDelegationSetLimitError
Errors returned by GetReusableDelegationSetLimit

GetTrafficPolicyError
Errors returned by GetTrafficPolicy

GetTrafficPolicyInstanceCountError
Errors returned by GetTrafficPolicyInstanceCount

GetTrafficPolicyInstanceError
Errors returned by GetTrafficPolicyInstance

ListGeoLocationsError
Errors returned by ListGeoLocations

ListHealthChecksError
Errors returned by ListHealthChecks

ListHostedZonesByNameError
Errors returned by ListHostedZonesByName

ListHostedZonesError
Errors returned by ListHostedZones

ListQueryLoggingConfigsError
Errors returned by ListQueryLoggingConfigs

ListResourceRecordSetsError
Errors returned by ListResourceRecordSets

ListReusableDelegationSetsError
Errors returned by ListReusableDelegationSets

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTagsForResourcesError
Errors returned by ListTagsForResources

ListTrafficPoliciesError
Errors returned by ListTrafficPolicies

ListTrafficPolicyInstancesByHostedZoneError
Errors returned by ListTrafficPolicyInstancesByHostedZone

ListTrafficPolicyInstancesByPolicyError
Errors returned by ListTrafficPolicyInstancesByPolicy

ListTrafficPolicyInstancesError
Errors returned by ListTrafficPolicyInstances

ListTrafficPolicyVersionsError
Errors returned by ListTrafficPolicyVersions

ListVPCAssociationAuthorizationsError
Errors returned by ListVPCAssociationAuthorizations

TestDNSAnswerError
Errors returned by TestDNSAnswer

UpdateHealthCheckError
Errors returned by UpdateHealthCheck

UpdateHostedZoneCommentError
Errors returned by UpdateHostedZoneComment

UpdateTrafficPolicyCommentError
Errors returned by UpdateTrafficPolicyComment

UpdateTrafficPolicyInstanceError
Errors returned by UpdateTrafficPolicyInstance

Traits

Route53
Trait representing the capabilities of the Route 53 API. Route 53 clients implement this trait.