Enum rusoto_route53::AssociateVPCWithHostedZoneError

pub enum AssociateVPCWithHostedZoneError {
    ConflictingDomainExists(String),
    InvalidInput(String),
    InvalidVPCId(String),
    LimitsExceeded(String),
    NoSuchHostedZone(String),
    NotAuthorized(String),
    PublicZoneVPCAssociation(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by AssociateVPCWithHostedZone

Variants
ConflictingDomainExists(String)

The cause of this error depends on whether you're trying to create a public or a private hosted zone:

Public hosted zone: Two hosted zones that have the same name or that have a parent/child relationship (example.com and test.example.com) can't have any common name servers. You tried to create a hosted zone that has the same name as an existing hosted zone or that's the parent or child of an existing hosted zone, and you specified a delegation set that shares one or more name servers with the existing hosted zone.

Private hosted zone: You specified an Amazon VPC that you're already using for another hosted zone, and the domain that you specified for one of the hosted zones is a subdomain of the domain that you specified for the other hosted zone. For example, you can't use the same Amazon VPC for the hosted zones for example.com and test.example.com.

InvalidInput(String)

The input is not valid.

InvalidVPCId(String)

The VPC ID that you specified either isn't a valid ID or the current account is not authorized to access this VPC.

LimitsExceeded(String)

The limits specified for a resource have been exceeded.

NoSuchHostedZone(String)

No hosted zone exists with the ID that you specified.

NotAuthorized(String)

Associating the specified VPC with the specified hosted zone has not been authorized.

PublicZoneVPCAssociation(String)

You're trying to associate a VPC with a public hosted zone. Amazon Route 53 doesn't support associating a VPC with a public hosted zone.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl AssociateVPCWithHostedZoneError

pub fn from_body(body: &str) -> AssociateVPCWithHostedZoneError