Enum rusoto_route53::CreateHostedZoneError

pub enum CreateHostedZoneError {
    ConflictingDomainExists(String),
    DelegationSetNotAvailable(String),
    DelegationSetNotReusable(String),
    HostedZoneAlreadyExists(String),
    InvalidDomainName(String),
    InvalidInput(String),
    InvalidVPCId(String),
    NoSuchDelegationSet(String),
    TooManyHostedZones(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateHostedZone

Variants
ConflictingDomainExists(String)

The cause of this error depends on whether you're trying to create a public or a private hosted zone:

Public hosted zone: Two hosted zones that have the same name or that have a parent/child relationship (example.com and test.example.com) can't have any common name servers. You tried to create a hosted zone that has the same name as an existing hosted zone or that's the parent or child of an existing hosted zone, and you specified a delegation set that shares one or more name servers with the existing hosted zone.

Private hosted zone: You specified an Amazon VPC that you're already using for another hosted zone, and the domain that you specified for one of the hosted zones is a subdomain of the domain that you specified for the other hosted zone. For example, you can't use the same Amazon VPC for the hosted zones for example.com and test.example.com.

DelegationSetNotAvailable(String)

You can create a hosted zone that has the same name as an existing hosted zone (example.com is common), but there is a limit to the number of hosted zones that have the same name. If you get this error, Amazon Route 53 has reached that limit. If you own the domain name and Amazon Route 53 generates this error, contact Customer Support.

DelegationSetNotReusable(String)

A reusable delegation set with the specified ID does not exist.

HostedZoneAlreadyExists(String)

The hosted zone you're trying to create already exists. Amazon Route 53 returns this error when a hosted zone has already been created with the specified CallerReference.

InvalidDomainName(String)

The specified domain name is not valid.

InvalidInput(String)

The input is not valid.

InvalidVPCId(String)

The VPC ID that you specified either isn't a valid ID or the current account is not authorized to access this VPC.

NoSuchDelegationSet(String)

A reusable delegation set with the specified ID does not exist.

TooManyHostedZones(String)

This hosted zone can't be created because the hosted zone limit is exceeded. To request a limit increase, go to the Amazon Route 53 Contact Us page.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateHostedZoneError

pub fn from_body(body: &str) -> CreateHostedZoneError