Enum rusoto_route53::CreateReusableDelegationSetError

pub enum CreateReusableDelegationSetError {
    DelegationSetAlreadyCreated(String),
    DelegationSetAlreadyReusable(String),
    DelegationSetNotAvailable(String),
    HostedZoneNotFound(String),
    InvalidArgument(String),
    InvalidInput(String),
    LimitsExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateReusableDelegationSet

Variants
DelegationSetAlreadyCreated(String)

A delegation set with the same owner and caller reference combination has already been created.

DelegationSetAlreadyReusable(String)

The specified delegation set has already been marked as reusable.

DelegationSetNotAvailable(String)

You can create a hosted zone that has the same name as an existing hosted zone (example.com is common), but there is a limit to the number of hosted zones that have the same name. If you get this error, Amazon Route 53 has reached that limit. If you own the domain name and Amazon Route 53 generates this error, contact Customer Support.

HostedZoneNotFound(String)

The specified HostedZone can't be found.

InvalidArgument(String)

Parameter name is invalid.

InvalidInput(String)

The input is not valid.

LimitsExceeded(String)

The limits specified for a resource have been exceeded.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateReusableDelegationSetError

fn from_body(body: &str) -> CreateReusableDelegationSetError