Enum rusoto_route53::CreateTrafficPolicyError

pub enum CreateTrafficPolicyError {
    InvalidInput(String),
    InvalidTrafficPolicyDocument(String),
    TooManyTrafficPolicies(String),
    TrafficPolicyAlreadyExists(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateTrafficPolicy

Variants
InvalidInput(String)

The input is not valid.

InvalidTrafficPolicyDocument(String)

The format of the traffic policy document that you specified in the Document element is invalid.

TooManyTrafficPolicies(String)

This traffic policy can't be created because the current account has reached the limit on the number of traffic policies.

For information about default limits, see Limits in the Amazon Route 53 Developer Guide.

To get the current limit for an account, see GetAccountLimit.

To request a higher limit, create a case with the AWS Support Center.

TrafficPolicyAlreadyExists(String)

A traffic policy that has the same value for Name already exists.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateTrafficPolicyError

pub fn from_body(body: &str) -> CreateTrafficPolicyError