Struct rusoto_route53::Change

pub struct Change {
    pub action: String,
    pub resource_record_set: ResourceRecordSet,
}

The information for each resource record set that you want to change.

Fields
action: String

The action to perform:

CREATE: Creates a resource record set that has the specified values.

DELETE: Deletes a existing resource record set.

To delete the resource record set that is associated with a traffic policy instance, use DeleteTrafficPolicyInstance . Amazon Route 53 will delete the resource record set automatically. If you delete the resource record set by using ChangeResourceRecordSets, Amazon Route 53 doesn't automatically delete the traffic policy instance, and you'll continue to be charged for it even though it's no longer in use.

UPSERT: If a resource record set doesn't already exist, Amazon Route 53 creates it. If a resource record set does exist, Amazon Route 53 updates it with the values in the request.

The values that you need to include in the request depend on the type of resource record set that you're creating, deleting, or updating:

Basic resource record sets (excluding alias, failover, geolocation, latency, and weighted resource record sets)

Name

Type

TTL

Failover, geolocation, latency, or weighted resource record sets (excluding alias resource record sets)

Name

Type

TTL

SetIdentifier

Alias resource record sets (including failover alias, geolocation alias, latency alias, and weighted alias resource record sets)

Name

Type

AliasTarget (includes DNSName, EvaluateTargetHealth, and HostedZoneId)

SetIdentifier (for failover, geolocation, latency, and weighted resource record sets)

resource_record_set: ResourceRecordSet

Information about the resource record set to create, delete, or update.