Struct rusoto_route53::CloudWatchAlarmConfiguration

pub struct CloudWatchAlarmConfiguration {
    pub comparison_operator: String,
    pub dimensions: Option<Vec<Dimension>>,
    pub evaluation_periods: i64,
    pub metric_name: String,
    pub namespace: String,
    pub period: i64,
    pub statistic: String,
    pub threshold: f64,
}

A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

Fields
comparison_operator: String

For the metric that the CloudWatch alarm is associated with, the arithmetic operation that is used for the comparison.

dimensions: Option<Vec<Dimension>>

For the metric that the CloudWatch alarm is associated with, a complex type that contains information about the dimensions for the metric. For information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

evaluation_periods: i64

For the metric that the CloudWatch alarm is associated with, the number of periods that the metric is compared to the threshold.

metric_name: String

The name of the CloudWatch metric that the alarm is associated with.

namespace: String

The namespace of the metric that the alarm is associated with. For more information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

period: i64

For the metric that the CloudWatch alarm is associated with, the duration of one evaluation period in seconds.

statistic: String

For the metric that the CloudWatch alarm is associated with, the statistic that is applied to the metric.

threshold: f64

For the metric that the CloudWatch alarm is associated with, the value the metric is compared with.