Struct rusoto_route53::CreateHostedZoneRequest

pub struct CreateHostedZoneRequest {
    pub caller_reference: String,
    pub delegation_set_id: Option<String>,
    pub hosted_zone_config: Option<HostedZoneConfig>,
    pub name: String,
    pub vpc: Option<VPC>,
}

A complex type that contains information about the request to create a hosted zone.

Fields
caller_reference: String

A unique string that identifies the request and that allows failed CreateHostedZone requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you submit a CreateHostedZone request. CallerReference can be any unique string, for example, a date/time stamp.

delegation_set_id: Option<String>

If you want to associate a reusable delegation set with this hosted zone, the ID that Amazon Route 53 assigned to the reusable delegation set when you created it. For more information about reusable delegation sets, see CreateReusableDelegationSet.

hosted_zone_config: Option<HostedZoneConfig>

(Optional) A complex type that contains the following optional values:

For public and private hosted zones, an optional comment

For private hosted zones, an optional PrivateZone element

If you don't specify a comment or the PrivateZone element, omit HostedZoneConfig and the other elements.

name: String

The name of the domain. For resource record types that include a domain name, specify a fully qualified domain name, for example, www.example.com. The trailing dot is optional; Amazon Route 53 assumes that the domain name is fully qualified. This means that Amazon Route 53 treats www.example.com (without a trailing dot) and www.example.com. (with a trailing dot) as identical.

If you're creating a public hosted zone, this is the name you have registered with your DNS registrar. If your domain name is registered with a registrar other than Amazon Route 53, change the name servers for your domain to the set of NameServers that CreateHostedZone returns in DelegationSet.

vpc: Option<VPC>

(Private hosted zones only) A complex type that contains information about the Amazon VPC that you're associating with this hosted zone.

You can specify only one Amazon VPC when you create a private hosted zone. To associate additional Amazon VPCs with the hosted zone, use AssociateVPCWithHostedZone after you create a hosted zone.