Struct rusoto_route53::ListGeoLocationsRequest

pub struct ListGeoLocationsRequest {
    pub max_items: Option<String>,
    pub start_continent_code: Option<String>,
    pub start_country_code: Option<String>,
    pub start_subdivision_code: Option<String>,
}

A request to get a list of geographic locations that Amazon Route 53 supports for geolocation resource record sets.

Fields
max_items: Option<String>

(Optional) The maximum number of geolocations to be included in the response body for this request. If more than MaxItems geolocations remain to be listed, then the value of the IsTruncated element in the response is true.

start_continent_code: Option<String>

The code for the continent with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Amazon Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextContinentCode from the previous response has a value, enter that value in StartContinentCode to return the next page of results.

Include StartContinentCode only if you want to list continents. Don't include StartContinentCode when you're listing countries or countries with their subdivisions.

start_country_code: Option<String>

The code for the country with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Amazon Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextCountryCode from the previous response has a value, enter that value in StartCountryCode to return the next page of results.

Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2.

start_subdivision_code: Option<String>

The code for the subdivision (for example, state or province) with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Amazon Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextSubdivisionCode from the previous response has a value, enter that value in StartSubdivisionCode to return the next page of results.

To list subdivisions of a country, you must include both StartCountryCode and StartSubdivisionCode.