Struct rusoto_route53::ListGeoLocationsResponse

pub struct ListGeoLocationsResponse {
    pub geo_location_details_list: Vec<GeoLocationDetails>,
    pub is_truncated: bool,
    pub max_items: String,
    pub next_continent_code: Option<String>,
    pub next_country_code: Option<String>,
    pub next_subdivision_code: Option<String>,
}

A complex type containing the response information for the request.

Fields
geo_location_details_list: Vec<GeoLocationDetails>

A complex type that contains one GeoLocationDetails element for each location that Amazon Route 53 supports for geolocation.

is_truncated: bool

A value that indicates whether more locations remain to be listed after the last location in this response. If so, the value of IsTruncated is true. To get more values, submit another request and include the values of NextContinentCode, NextCountryCode, and NextSubdivisionCode in the StartContinentCode, StartCountryCode, and StartSubdivisionCode, as applicable.

max_items: String

The value that you specified for MaxItems in the request.

next_continent_code: Option<String>

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextContinentCode in the StartContinentCode parameter in another ListGeoLocations request.

next_country_code: Option<String>

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextCountryCode in the StartCountryCode parameter in another ListGeoLocations request.

next_subdivision_code: Option<String>

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextSubdivisionCode in the StartSubdivisionCode parameter in another ListGeoLocations request.