Struct rusoto_route53::ListTrafficPolicyInstancesByHostedZoneRequest

pub struct ListTrafficPolicyInstancesByHostedZoneRequest {
    pub hosted_zone_id: String,
    pub max_items: Option<String>,
    pub traffic_policy_instance_name_marker: Option<String>,
    pub traffic_policy_instance_type_marker: Option<String>,
}

A request for the traffic policy instances that you created in a specified hosted zone.

Fields
hosted_zone_id: String

The ID of the hosted zone that you want to list traffic policy instances for.

max_items: Option<String>

The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true, and the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

traffic_policy_instance_name_marker: Option<String>

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename, specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

traffic_policy_instance_type_marker: Option<String>

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype, specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.