Struct rusoto_route53::ListTrafficPolicyVersionsResponse

pub struct ListTrafficPolicyVersionsResponse {
    pub is_truncated: bool,
    pub max_items: String,
    pub traffic_policies: Vec<TrafficPolicy>,
    pub traffic_policy_version_marker: String,
}

A complex type that contains the response information for the request.

Fields
is_truncated: bool

A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicyVersions request and specifying the value of NextMarker in the marker parameter.

max_items: String

The value that you specified for the maxitems parameter in the ListTrafficPolicyVersions request that produced the current response.

traffic_policies: Vec<TrafficPolicy>

A list that contains one TrafficPolicy element for each traffic policy version that is associated with the specified traffic policy.

traffic_policy_version_marker: String

If IsTruncated is true, the value of TrafficPolicyVersionMarker identifies the first traffic policy that Amazon Route 53 will return if you submit another request. Call ListTrafficPolicyVersions again and specify the value of TrafficPolicyVersionMarker in the TrafficPolicyVersionMarker request parameter.

This element is present only if IsTruncated is true.