Struct rusoto_route53::Route53Client

pub struct Route53Client { /* fields omitted */ }

A client for the Route 53 API.

Methods

impl Route53Client

pub fn new(region: Region) -> Route53Client

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> Route53Client where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,