Struct rusoto_route53::TrafficPolicyInstance

pub struct TrafficPolicyInstance {
    pub hosted_zone_id: String,
    pub id: String,
    pub message: String,
    pub name: String,
    pub state: String,
    pub ttl: i64,
    pub traffic_policy_id: String,
    pub traffic_policy_type: String,
    pub traffic_policy_version: i64,
}

A complex type that contains settings for the new traffic policy instance.

Fields
hosted_zone_id: String

The ID of the hosted zone that Amazon Route 53 created resource record sets in.

id: String

The ID that Amazon Route 53 assigned to the new traffic policy instance.

message: String

If State is Failed, an explanation of the reason for the failure. If State is another value, Message is empty.

name: String

The DNS name, such as www.example.com, for which Amazon Route 53 responds to queries by using the resource record sets that are associated with this traffic policy instance.

state: String

The value of State is one of the following values:

Applied

Amazon Route 53 has finished creating resource record sets, and changes have propagated to all Amazon Route 53 edge locations.

Creating

Amazon Route 53 is creating the resource record sets. Use GetTrafficPolicyInstance to confirm that the CreateTrafficPolicyInstance request completed successfully.

Failed

Amazon Route 53 wasn't able to create or update the resource record sets. When the value of State is Failed, see Message for an explanation of what caused the request to fail.

ttl: i64

The TTL that Amazon Route 53 assigned to all of the resource record sets that it created in the specified hosted zone.

traffic_policy_id: String

The ID of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.

traffic_policy_type: String

The DNS type that Amazon Route 53 assigned to all of the resource record sets that it created for this traffic policy instance.

traffic_policy_version: i64

The version of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.